feat: make modern Print Format Builder the default; move old to classic
- Rename print_format_builder_beta page → print-format-builder (drops "beta")
- Rename old print_format_builder page → print-format-builder-classic
- create_custom_format defaults beta=True
- Customize flows in print.js create beta formats and route to new builder
- Classic builder's show_start redirects to new builder (no create entry)
- Patch removes the orphan print-format-builder-beta Page record
