+++ frappe/printing/page/print_format_builder/print_format_builder.json
@@ -1,23 +1,22 @@
 {
- "content": null, 
- "creation": "2015-01-27 04:35:43.872918", 
- "docstatus": 0, 
- "doctype": "Page", 
- "idx": 0, 
- "modified": "2017-05-03 05:59:33.702308", 
- "modified_by": "Administrator", 
- "module": "Printing", 
- "name": "print-format-builder", 
- "owner": "Administrator", 
- "page_name": "print-format-builder", 
+ "content": null,
+ "creation": "2021-07-10 12:22:16.138485",
+ "docstatus": 0,
+ "doctype": "Page",
+ "idx": 0,
+ "modified": "2021-07-10 12:22:16.138485",
+ "modified_by": "Administrator",
+ "module": "Printing",
+ "name": "print-format-builder",
+ "owner": "Administrator",
+ "page_name": "Print Format Builder",
  "roles": [
   {
    "role": "System Manager"
   }
- ], 
- "script": null, 
- "standard": "Yes", 
- "style": null, 
- "system_page": 0, 
- "title": "Print Format Builder"
+ ],
+ "script": null,
+ "standard": "Yes",
+ "style": null,
+ "system_page": 0
 }
