--- frappe/printing/page/print_format_builder_beta/print_format_builder_beta.css
@@ -1,3 +1,0 @@
-.layout-main-section-wrapper {
-	margin-bottom: 0;
-}
