+++ frappe/printing/page/print_format_builder_classic/print_format_builder_classic.py
@@ -1,0 +1,23 @@
+import frappe
+
+
+@frappe.whitelist()
+def create_custom_format(
+	doctype: str, name: str | int, based_on: str = "Standard", beta: str | int | bool = True
+):
+	# beta defaults to True so new formats use the modern print format builder;
+	# pass beta=False explicitly to create a classic-builder format.
+	doc = frappe.new_doc("Print Format")
+	doc.doc_type = doctype
+	doc.name = name
+	beta = frappe.parse_json(beta)
+
+	if beta:
+		doc.print_format_builder_beta = 1
+	else:
+		doc.print_format_builder = 1
+	doc.format_data = (
+		frappe.db.get_value("Print Format", based_on, "format_data") if based_on != "Standard" else None
+	)
+	doc.insert()
+	return doc
